rwicc: Regression with Interval-Censored Covariates

Provides functions to simulate and analyze data for a regression model with an interval censored covariate, as described in Morrison et al. (2021) doi:10.1111/biom.13472 .

The rwicc package implements a regression model with an interval-censored covariate using an EM algorithm, as described in Morrison et al (2021); doi:10.1111/biom.13472 .

rwicc functions

The main rwicc functions are:

simulate_interval_censoring

fit_joint_model

References

Morrison, Laeyendecker, and Brookmeyer (2021). "Regression with interval-censored covariates: Application to cross-sectional incidence estimation". Biometrics. doi:10.1111/biom.13472 .

See also

Author

Maintainer: Douglas Ezra Morrison demorrison@ucdavis.edu (ORCID) [copyright holder]

Authors:

Ron Brookmeyer